+++ include/designer/libabstractresourcebrowser.zig
@@ -3628,9 +3628,9 @@ pub const qdesignerresourcebrowserinterface = struct {
     ///
     /// ## Parameter(s):
     ///
-    /// ` param1: u64 `
+    /// ` param1: usize `
     ///
-    pub fn Find(param1: u64) QtC.QWidget {
+    pub fn Find(param1: usize) QtC.QWidget {
         return qtc.QWidget_Find(@bitCast(param1));
     }
