+++ include/extras-kcoreaddons/libkcompositejob.zig
@@ -678,7 +678,7 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    pub fn ProcessedAmount(self: ?*anyopaque, unit: i32) u64 {
+    pub fn ProcessedAmount(self: ?*anyopaque, unit: i32) usize {
         return qtc.KJob_ProcessedAmount(@ptrCast(self), @bitCast(unit));
     }
 
@@ -692,7 +692,7 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    pub fn TotalAmount(self: ?*anyopaque, unit: i32) u64 {
+    pub fn TotalAmount(self: ?*anyopaque, unit: i32) usize {
         return qtc.KJob_TotalAmount(@ptrCast(self), @bitCast(unit));
     }
 
@@ -704,7 +704,7 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    pub fn Percent(self: ?*anyopaque) u64 {
+    pub fn Percent(self: ?*anyopaque) usize {
         return qtc.KJob_Percent(@ptrCast(self));
     }
 
@@ -860,9 +860,9 @@ pub const kcompositejob = struct {
     ///
     /// ` job: QtC.KJob `
     ///
-    /// ` size: u64 `
+    /// ` size: usize `
     ///
-    pub fn TotalSize(self: ?*anyopaque, job: ?*anyopaque, size: u64) void {
+    pub fn TotalSize(self: ?*anyopaque, job: ?*anyopaque, size: usize) void {
         qtc.KJob_TotalSize(@ptrCast(self), @ptrCast(job), @bitCast(size));
     }
 
@@ -874,9 +874,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, size: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, size: usize) callconv(.c) void `
     ///
-    pub fn OnTotalSize(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnTotalSize(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, usize) callconv(.c) void) void {
         qtc.KJob_Connect_TotalSize(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -890,9 +890,9 @@ pub const kcompositejob = struct {
     ///
     /// ` job: QtC.KJob `
     ///
-    /// ` size: u64 `
+    /// ` size: usize `
     ///
-    pub fn ProcessedSize(self: ?*anyopaque, job: ?*anyopaque, size: u64) void {
+    pub fn ProcessedSize(self: ?*anyopaque, job: ?*anyopaque, size: usize) void {
         qtc.KJob_ProcessedSize(@ptrCast(self), @ptrCast(job), @bitCast(size));
     }
 
@@ -904,9 +904,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, size: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, size: usize) callconv(.c) void `
     ///
-    pub fn OnProcessedSize(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnProcessedSize(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, usize) callconv(.c) void) void {
         qtc.KJob_Connect_ProcessedSize(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -920,9 +920,9 @@ pub const kcompositejob = struct {
     ///
     /// ` job: QtC.KJob `
     ///
-    /// ` speed: u64 `
+    /// ` speed: usize `
     ///
-    pub fn Speed(self: ?*anyopaque, job: ?*anyopaque, speed: u64) void {
+    pub fn Speed(self: ?*anyopaque, job: ?*anyopaque, speed: usize) void {
         qtc.KJob_Speed(@ptrCast(self), @ptrCast(job), @bitCast(speed));
     }
 
@@ -934,9 +934,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, speed: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, speed: usize) callconv(.c) void `
     ///
-    pub fn OnSpeed(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnSpeed(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, usize) callconv(.c) void) void {
         qtc.KJob_Connect_Speed(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -2552,9 +2552,9 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    /// ` amount: u64 `
+    /// ` amount: usize `
     ///
-    pub fn SetProcessedAmount(self: ?*anyopaque, unit: i32, amount: u64) void {
+    pub fn SetProcessedAmount(self: ?*anyopaque, unit: i32, amount: usize) void {
         qtc.KCompositeJob_SetProcessedAmount(@ptrCast(self), @bitCast(unit), @bitCast(amount));
     }
 
@@ -2574,9 +2574,9 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    /// ` amount: u64 `
+    /// ` amount: usize `
     ///
-    pub fn SuperSetProcessedAmount(self: ?*anyopaque, unit: i32, amount: u64) void {
+    pub fn SuperSetProcessedAmount(self: ?*anyopaque, unit: i32, amount: usize) void {
         qtc.KCompositeJob_SuperSetProcessedAmount(@ptrCast(self), @bitCast(unit), @bitCast(amount));
     }
 
@@ -2590,9 +2590,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob`
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, unit: kjob_enums.Unit, amount: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, unit: kjob_enums.Unit, amount: usize) callconv(.c) void `
     ///
-    pub fn OnSetProcessedAmount(self: ?*anyopaque, callback: *const fn (?*anyopaque, i32, u64) callconv(.c) void) void {
+    pub fn OnSetProcessedAmount(self: ?*anyopaque, callback: *const fn (?*anyopaque, i32, usize) callconv(.c) void) void {
         qtc.KCompositeJob_OnSetProcessedAmount(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -2608,9 +2608,9 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    /// ` amount: u64 `
+    /// ` amount: usize `
     ///
-    pub fn SetTotalAmount(self: ?*anyopaque, unit: i32, amount: u64) void {
+    pub fn SetTotalAmount(self: ?*anyopaque, unit: i32, amount: usize) void {
         qtc.KCompositeJob_SetTotalAmount(@ptrCast(self), @bitCast(unit), @bitCast(amount));
     }
 
@@ -2630,9 +2630,9 @@ pub const kcompositejob = struct {
     ///
     /// ` unit: kjob_enums.Unit `
     ///
-    /// ` amount: u64 `
+    /// ` amount: usize `
     ///
-    pub fn SuperSetTotalAmount(self: ?*anyopaque, unit: i32, amount: u64) void {
+    pub fn SuperSetTotalAmount(self: ?*anyopaque, unit: i32, amount: usize) void {
         qtc.KCompositeJob_SuperSetTotalAmount(@ptrCast(self), @bitCast(unit), @bitCast(amount));
     }
 
@@ -2646,9 +2646,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob`
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, unit: kjob_enums.Unit, amount: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, unit: kjob_enums.Unit, amount: usize) callconv(.c) void `
     ///
-    pub fn OnSetTotalAmount(self: ?*anyopaque, callback: *const fn (?*anyopaque, i32, u64) callconv(.c) void) void {
+    pub fn OnSetTotalAmount(self: ?*anyopaque, callback: *const fn (?*anyopaque, i32, usize) callconv(.c) void) void {
         qtc.KCompositeJob_OnSetTotalAmount(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -2714,9 +2714,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` percentage: u64 `
+    /// ` percentage: usize `
     ///
-    pub fn SetPercent(self: ?*anyopaque, percentage: u64) void {
+    pub fn SetPercent(self: ?*anyopaque, percentage: usize) void {
         qtc.KCompositeJob_SetPercent(@ptrCast(self), @bitCast(percentage));
     }
 
@@ -2734,9 +2734,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` percentage: u64 `
+    /// ` percentage: usize `
     ///
-    pub fn SuperSetPercent(self: ?*anyopaque, percentage: u64) void {
+    pub fn SuperSetPercent(self: ?*anyopaque, percentage: usize) void {
         qtc.KCompositeJob_SuperSetPercent(@ptrCast(self), @bitCast(percentage));
     }
 
@@ -2750,9 +2750,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob`
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, percentage: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, percentage: usize) callconv(.c) void `
     ///
-    pub fn OnSetPercent(self: ?*anyopaque, callback: *const fn (?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnSetPercent(self: ?*anyopaque, callback: *const fn (?*anyopaque, usize) callconv(.c) void) void {
         qtc.KCompositeJob_OnSetPercent(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -2814,11 +2814,11 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` processedAmount: u64 `
+    /// ` processedAmount: usize `
     ///
-    /// ` totalAmount: u64 `
+    /// ` totalAmount: usize `
     ///
-    pub fn EmitPercent(self: ?*anyopaque, processedAmount: u64, totalAmount: u64) void {
+    pub fn EmitPercent(self: ?*anyopaque, processedAmount: usize, totalAmount: usize) void {
         qtc.KCompositeJob_EmitPercent(@ptrCast(self), @bitCast(processedAmount), @bitCast(totalAmount));
     }
 
@@ -2836,11 +2836,11 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` processedAmount: u64 `
+    /// ` processedAmount: usize `
     ///
-    /// ` totalAmount: u64 `
+    /// ` totalAmount: usize `
     ///
-    pub fn SuperEmitPercent(self: ?*anyopaque, processedAmount: u64, totalAmount: u64) void {
+    pub fn SuperEmitPercent(self: ?*anyopaque, processedAmount: usize, totalAmount: usize) void {
         qtc.KCompositeJob_SuperEmitPercent(@ptrCast(self), @bitCast(processedAmount), @bitCast(totalAmount));
     }
 
@@ -2854,9 +2854,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob`
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, processedAmount: u64, totalAmount: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, processedAmount: usize, totalAmount: usize) callconv(.c) void `
     ///
-    pub fn OnEmitPercent(self: ?*anyopaque, callback: *const fn (?*anyopaque, u64, u64) callconv(.c) void) void {
+    pub fn OnEmitPercent(self: ?*anyopaque, callback: *const fn (?*anyopaque, usize, usize) callconv(.c) void) void {
         qtc.KCompositeJob_OnEmitPercent(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -2870,9 +2870,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` speed: u64 `
+    /// ` speed: usize `
     ///
-    pub fn EmitSpeed(self: ?*anyopaque, speed: u64) void {
+    pub fn EmitSpeed(self: ?*anyopaque, speed: usize) void {
         qtc.KCompositeJob_EmitSpeed(@ptrCast(self), @bitCast(speed));
     }
 
@@ -2890,9 +2890,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` speed: u64 `
+    /// ` speed: usize `
     ///
-    pub fn SuperEmitSpeed(self: ?*anyopaque, speed: u64) void {
+    pub fn SuperEmitSpeed(self: ?*anyopaque, speed: usize) void {
         qtc.KCompositeJob_SuperEmitSpeed(@ptrCast(self), @bitCast(speed));
     }
 
@@ -2906,9 +2906,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob`
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, speed: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, speed: usize) callconv(.c) void `
     ///
-    pub fn OnEmitSpeed(self: ?*anyopaque, callback: *const fn (?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnEmitSpeed(self: ?*anyopaque, callback: *const fn (?*anyopaque, usize) callconv(.c) void) void {
         qtc.KCompositeJob_OnEmitSpeed(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -3236,9 +3236,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, unit: kjob_enums.Unit, amount: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, unit: kjob_enums.Unit, amount: usize) callconv(.c) void `
     ///
-    pub fn OnTotalAmountChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, i32, u64) callconv(.c) void) void {
+    pub fn OnTotalAmountChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, i32, usize) callconv(.c) void) void {
         qtc.KJob_Connect_TotalAmountChanged(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -3252,9 +3252,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, unit: kjob_enums.Unit, amount: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, unit: kjob_enums.Unit, amount: usize) callconv(.c) void `
     ///
-    pub fn OnProcessedAmountChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, i32, u64) callconv(.c) void) void {
+    pub fn OnProcessedAmountChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, i32, usize) callconv(.c) void) void {
         qtc.KJob_Connect_ProcessedAmountChanged(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
 
@@ -3268,9 +3268,9 @@ pub const kcompositejob = struct {
     ///
     /// ` self: QtC.KCompositeJob `
     ///
-    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, percent: u64) callconv(.c) void `
+    /// ` callback: *const fn (self: QtC.KCompositeJob, job: QtC.KJob, percent: usize) callconv(.c) void `
     ///
-    pub fn OnPercentChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, u64) callconv(.c) void) void {
+    pub fn OnPercentChanged(self: ?*anyopaque, callback: *const fn (?*anyopaque, ?*anyopaque, usize) callconv(.c) void) void {
         qtc.KJob_Connect_PercentChanged(@ptrCast(self), @bitCast(@intFromPtr(callback)));
     }
